fix warnings

+++ src/ssphh_ppmcompare.cpp
@@ -199,9 +199,9 @@ namespace SSPHH
 			const int AvgIIndex = 4;
 			const int MinIIndex = 5;
 			const int MaxIIndex = 6;
-			for (int j = 0; j < blockstat.image.height(); j++) {
-				for (int i = 0; i < blockstat.image.width(); i++) {
-					size_t addr = blockstat.image.addr(i, j);
+			for (unsigned j = 0; j < blockstat.image.height(); j++) {
+				for (unsigned i = 0; i < blockstat.image.width(); i++) {
+					unsigned addr = blockstat.image.addr(i, j);
 					float averageI = std::get<AvgIIndex>(blockcounts[addr]);
 					float mindiffI = std::get<MinIIndex>(blockcounts[addr]);
 					float maxdiffI = std::get<MaxIIndex>(blockcounts[addr]);
